Made edits for the text entries
I was able to edit the majority of all text entry boxes for all users on the site. A couple small tweaks still need to be made for the trainer add/edit notes, and possibly the address field wherever it is located. Need to tweak the sign up page for things to line up.